Move `is_deferred_callable` to util/deferred_call.hpp
